# NavyRadar.Shared/Domain/RegisterDto.cs
namespace NavyRadar.Shared.Domain;

public class RegisterDto
{
    public required string Username { get; set; }
    public required string Email { get; set; }
    public required string Password { get; set; }
}